requirements.txt
numpy==1.26.4
pandas==2.2.1
requests==2.31.0
joblib==1.3.2
textblob==0.17.1
python-dotenv==1.0.1
tensorflow==2.18.0
ml-dtypes<0.5.0,>=0.4.0
tensorboard<2.19,>=2.18
scikit-learn==1.4.1.post1
fastapi==0.110.0
uvicorn==0.29.0
alpaca-trade-api